+++ feature-mypage/src/main/java/com/materip/feature_mypage/view_models/MyPage/ProfileDescriptionViewModel.kt
@@ -1,4 +1,4 @@
-package com.materip.feature_mypage.view_models
+package com.materip.feature_mypage.view_models.MyPage
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
@@ -6,9 +6,8 @@ import androidx.lifecycle.viewModelScope
 import com.materip.core_common.ErrorState
 import com.materip.core_common.Result
 import com.materip.core_common.asResult
-import com.materip.core_model.response.GetProfileResponseDto
-import com.materip.core_repository.repository.profile_repository.ProfileRepository
-import com.materip.core_repository.useCase.GetProfileUseCase
+import com.materip.core_model.response.GetProfileDetailsResponseDto
+import com.materip.core_repository.useCase.GetProfileDetails
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
@@ -20,13 +19,13 @@ import kotlinx.coroutines.flow.update
 import javax.inject.Inject
 
 @HiltViewModel
-class ProfileMainViewModel @Inject constructor(
-    private val getProfileUseCase: GetProfileUseCase
+class ProfileDescriptionViewModel @Inject constructor(
+    private val getProfileDetails: GetProfileDetails
 ): ViewModel() {
     private val invalidTokenError = MutableStateFlow<Boolean>(false)
     private val notFoundTokenError = MutableStateFlow<Boolean>(false)
     private val generalError = MutableStateFlow<Pair<Boolean, String?>>(Pair(false, null))
-    val errorState: StateFlow<ErrorState> = combine(invalidTokenError, notFoundTokenError, generalError){invalidToken, notFoundToken, general ->
+    val errorState: StateFlow<ErrorState> = combine(invalidTokenError, notFoundTokenError, generalError){ invalidToken, notFoundToken, general ->
         ErrorState.AuthError(
             invalidToken,
             notFoundToken,
@@ -37,39 +36,38 @@ class ProfileMainViewModel @Inject constructor(
         started = SharingStarted.WhileSubscribed(5_000),
         initialValue = ErrorState.Loading
     )
-    val uiState: StateFlow<ProfileMainUiState> = errorState.map {
-        if (it is ErrorState.AuthError && it.isInvalid()){throw Exception("")}
-        val result = getProfileUseCase()
+    val uiState: StateFlow<ProfileDescriptionUiState> = errorState.map{
+        if(it is ErrorState.AuthError && it.isInvalid()) throw Exception("")
+        Log.d("MATETRIP API TEST", "result before")
+        val result = getProfileDetails()
+        Log.d("MATETRIP API TEST", "result : ${result}")
         if (result.error != null){
             when(result.error!!.status){
-                401 -> notFoundTokenError.update{true}
-                404 -> invalidTokenError.update{true}
+                404 -> notFoundTokenError.update{true}
                 else -> generalError.update{Pair(true, result.error!!.message)}
             }
-            throw Exception("Error occur in Get Profile")
+            throw Exception("")
         }
         result.data
     }.asResult().map{result ->
+        Log.d("MATETRIP API TEST", "map result : ${result}")
         when(result){
-            Result.Loading -> ProfileMainUiState.Loading
-            is Result.Success -> ProfileMainUiState.Success(result.data!!)
-            is Result.Error -> ProfileMainUiState.Error
+            Result.Loading -> ProfileDescriptionUiState.Loading
+            is Result.Success -> ProfileDescriptionUiState.Success(user = result.data!!)
+            is Result.Error -> ProfileDescriptionUiState.Error
         }
     }.stateIn(
         scope = viewModelScope,
         started = SharingStarted.WhileSubscribed(5_000),
-        initialValue = ProfileMainUiState.Loading
+        initialValue = ProfileDescriptionUiState.Loading
     )
-
 }
 
-sealed interface ProfileMainUiState{
+sealed interface ProfileDescriptionUiState{
+    data object Loading: ProfileDescriptionUiState
     data class Success(
-        val user: GetProfileResponseDto
-    ): ProfileMainUiState {
-        fun getTags(): List<String>{
-            return user.travelPreferences.plus(user.travelStyles).plus(user.foodPreferences)
-        }
+        val user: GetProfileDetailsResponseDto
+    ): ProfileDescriptionUiState {
         fun getAge(): String{
             return when(user.birthYear){
                 in 0..9 -> "${user.birthYear}세"
@@ -85,6 +83,5 @@ sealed interface ProfileMainUiState{
             }
         }
     }
-    data object Loading: ProfileMainUiState
-    data object Error: ProfileMainUiState
+    data object Error: ProfileDescriptionUiState
 }
